fix: わかりにくかったディレクトリ構成を修正
