+++ src/app/page.tsx
@@ -1,7 +1,7 @@
-import Header from "@/components/organisms/header";
-import Cards from "@/components/organisms/cards";
-import Footer from "@/components/organisms/footer";
-import TopHeader from "@/components/organisms/top-header";
+import Header from "@/components/Home/Header";
+import Cards from "@/components/Home/Cards";
+import Footer from "@/components/Home/Footer";
+import TopHeader from "@/components/Home/TopHeader";
 
 export default function Home() {
   return (
