+++ src/components/Home/Cards/AboutCard/about-card.module.scss
