+++ src/components/Home/Cards/AboutCard/index.tsx
