+++ src/components/Home/Cards/SkillCard/index.tsx
