+++ src/components/Home/Cards/index.tsx
@@ -1,0 +1,15 @@
+import AboutCard from "@/components/Home/Cards/AboutCard";
+import SkillCard from "@/components/Home/Cards/SkillCard";
+import ProjectCard from "@/components/Home/Cards/ProjectCard";
+
+const Cards = () => {
+  return (
+    <div>
+      <AboutCard />
+      <ProjectCard />
+      <SkillCard />
+    </div>
+  );
+};
+
+export default Cards;
