+++ src/components/Product/Detail/product-detail-template.module.scss
