--- src/components/organisms/cards/index.tsx
@@ -1,15 +1,0 @@
-import AboutCard from "@/components/molecules/card/about-card";
-import SkillCard from "@/components/molecules/card/skill-card";
-import ProjectCard from "@/components/molecules/card/project-card";
-
-const Cards = () => {
-  return (
-    <div>
-      <AboutCard />
-      <ProjectCard />
-      <SkillCard />
-    </div>
-  );
-};
-
-export default Cards;
